+++ lib/character_set/core_ext/regexp_ext.rb
@@ -4,6 +4,14 @@ class CharacterSet
       def character_set
         CharacterSet.of_regexp(self)
       end
+
+      def covered_by_character_set?(other)
+        other.superset?(character_set)
+      end
+
+      def uses_character_set?(other)
+        other.intersect?(character_set)
+      end
     end
   end
 end